specific use of float in devision function only

+++ calculator.c
@@ -2,26 +2,26 @@
  #include<string.h>
 int main()
 {
-    float n , m;
+    int n , m;
     char str[10];
 
     printf("choose numbers: ");
-    scanf("%f %f", &n , &m);
+    scanf("%d %d", &n , &m);
     
     printf("what do u wanna do?");
     scanf("%s" , &str);
 
     if (strcmp(str, "sum") == 0)  // Use strcmp() for string comparison
-    printf("Sum: %f" , n + m );
+    printf("Sum: %d" , n + m );
 
     else if (strcmp(str , "sub") == 0)
-    printf("Difference: %f" , n - m); 
+    printf("Difference: %d" , n - m); 
 
     else if (strcmp(str , "product") == 0)
-    printf("Product: %f" , n * m);
+    printf("Product: %d" , n * m);
 
     else if (strcmp(str , "devide") == 0)
-    printf("Quotient: %f" , n / m);
+    printf("Quotient: %.2f" , (float) n / m);
 
     else 
     printf("invalid operation\n");
